+++ main.cpp
@@ -2,8 +2,11 @@
 // Created by חי כרמי on 12/02/2024.
 #include <string>
 #include <iostream>
+#include "Field_manager.h"
 #include "MyMain.h"
+using namespace std;
 
+ int Field_manager::counter = 0;
 int main() {
     MyMain M1;
     return 0;